Move benchmarks into separate file
